+++ include/compose.bash
@@ -40,20 +40,7 @@ create-logfile() {
 }
 
 compose-up() {
-    declare desc="Starts containers with docker-compose"
-    declare services="$@"
-
-    deployer-generate
-
-    create-logfile
-
-    dockerCompose up -d $services
-
-    info "CloudBreak containers are started ..."
-    info "In a couple of minutes you can reach the UI (called Uluwatu)"
-    echo "  $ULU_HOST_ADDRESS" | blue
-    warn "Credentials are not printed here. You can get them by:"
-    echo '  cbd env show|grep "UAA_DEFAULT_USER_PW\|UAA_DEFAULT_USER_EMAIL"' | blue
+    dockerCompose up -d "$@"
 }
 
 compose-kill() {